# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Correspondence Workspace API",
    "description": "Built-in mailbox, event-driven case automation, connected system workspaces and human review.",
    "version": "0.9.0"
  },
  "paths": {
...
    "/api/simulator/tools": {
      "get": {
        "tags": [
          "Simulated systems"
        ],
        "summary": "Tools Registry",
        "operationId": "tools_registry_api_simulator_tools_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ToolDefinition"
                  },
                  "type": "array",
                  "title": "Response Tools Registry Api Simulator Tools Get"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "501": {
            "description": "Not Implemented",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/simulations/{simulation_id}/cases/{case_id}/tools": {
      "post": {
        "tags": [
          "Simulated systems"
        ],
        "summary": "Execute Tool",
        "operationId": "execute_tool_api_simulations__simulation_id__cases__case_id__tools_post",
        "parameters": [
          {
            "name": "simulation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Simulation Id"
            }
          },
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Case Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ToolRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ToolResult"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Implemented"
          }
        }
      }
    },
    "/api/simulations/{simulation_id}/cases/{case_id}/packages/{package_id}/file": {
      "get": {
        "tags": [
          "Simulated systems"
        ],
        "summary": "Package File",
        "operationId": "package_file_api_simulations__simulation_id__cases__case_id__packages__package_id__file_get",
        "parameters": [
          {
            "name": "simulation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Simulation Id"
            }
          },
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Case Id"
            }
          },
          {
            "name": "package_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Package Id"
            }
          },
          {
            "name": "loan_identifier",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Loan Identifier"
            }
          },
          {
            "name": "client_code",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Client Code"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Implemented"
          }
        }
      }
    },
    "/api/cases/{case_id}/workflow": {
      "get": {
        "tags": [
          "Case workflow"
        ],
        "summary": "Read Workflow",
        "operationId": "read_workflow_api_cases__case_id__workflow_get",
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Case Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowRead"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Implemented"
          }
        }
      }
    },
    "/api/cases/{case_id}/inputs": {
      "post": {
        "tags": [
          "Case workflow"
        ],
        "summary": "Supply Input",
        "operationId": "supply_input_api_cases__case_id__inputs_post",
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Case Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CaseInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseRead"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Implemented"
          }
        }
      }
    },
    "/api/cases/{case_id}/draft-edits": {
      "post": {
        "tags": [
          "Case workflow"
        ],
        "summary": "Edit Draft",
        "operationId": "edit_draft_api_cases__case_id__draft_edits_post",
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Case Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DraftEdit"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Edit Draft Api Cases  Case Id  Draft Edits Post"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Implemented"
          }
        }
      }
    },
    "/api/cases/{case_id}/handoff-acknowledgments": {
      "post": {
        "tags": [
          "Case workflow"
        ],
        "summary": "Acknowledge Handoff",
        "operationId": "acknowledge_handoff_api_cases__case_id__handoff_acknowledgments_post",
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Case Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HandoffAcknowledge"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Acknowledge Handoff Api Cases  Case Id  Handoff Acknowledgments Post"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Implemented"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CaseFamily": {
        "type": "string",
        "enum": [
          "profile_change",
          "document_request",
          "tax_payment",
          "credit_reporting",
          "loan_servicing"
        ],
        "title": "CaseFamily"
      },
      "CaseInput": {
        "properties": {
          "request_id": {
            "type": "string",
            "format": "uuid",
            "title": "Request Id"
          },
          "expected_revision": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Expected Revision"
          },
          "actor": {
            "type": "string",
            "maxLength": 254,
            "minLength": 1,
            "title": "Actor",
            "default": "Demo reviewer"
          },
          "kind": {
            "type": "string",
            "enum": [
              "name_legal_document",
              "tax_specialist_result",
              "bankruptcy_specialist_result",
              "eft_clarification",
              "amortization_document",
              "authority_missing",
              "authorized_representative",
              "assignment_exception",
              "resolve_assignment",
              "additional_concern",
              "unrelated_contact"
            ],
            "title": "Kind"
          },
          "text": {
            "type": "string",
            "maxLength": 12000,
            "title": "Text",
            "default": ""
          },
          "eft_intent": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "incoming_payment",
                  "outgoing_refund",
                  "heloc_draw"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Eft Intent"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "request_id",
          "expected_revision",
          "kind"
        ],
        "title": "CaseInput"
      },
...
      "CaseRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "simulation_id": {
            "type": "string",
            "format": "uuid",
            "title": "Simulation Id"
          },
          "loan_id": {
            "type": "string",
            "format": "uuid",
            "title": "Loan Id"
          },
          "loan_identifier": {
            "type": "string",
            "title": "Loan Identifier"
          },
          "ccid": {
            "type": "string",
            "title": "Ccid"
          },
          "client_code": {
            "type": "string",
            "title": "Client Code"
          },
          "borrower_display_name": {
            "type": "string",
            "title": "Borrower Display Name"
          },
          "balance_minor": {
            "type": "integer",
            "title": "Balance Minor"
          },
          "currency": {
            "type": "string",
            "title": "Currency"
          },
          "subject": {
            "type": "string",
            "title": "Subject"
          },
          "correspondence_text": {
            "type": "string",
            "title": "Correspondence Text"
          },
          "family": {
            "$ref": "#/components/schemas/CaseFamily"
          },
          "owner": {
            "type": "string",
            "title": "Owner"
          },
          "status": {
            "$ref": "#/components/schemas/CaseStatus"
          },
          "original_received_at": {
            "type": "string",
            "format": "date-time",
            "title": "Original Received At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "revision": {
            "type": "integer",
            "title": "Revision"
          },
          "classification": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Classification"
          },
          "content_revision": {
            "type": "integer",
            "title": "Content Revision"
          },
          "routing": {
            "additionalProperties": true,
            "type": "object",
            "title": "Routing"
          },
          "pending_work": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Pending Work"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "simulation_id",
          "loan_id",
          "loan_identifier",
          "ccid",
          "client_code",
          "borrower_display_name",
          "balance_minor",
          "currency",
          "subject",
          "correspondence_text",
          "family",
          "owner",
          "status",
          "original_received_at",
          "created_at",
          "updated_at",
          "revision",
          "classification",
          "content_revision",
          "routing",
          "pending_work"
        ],
        "title": "CaseRead"
      },
...
      "DraftEdit": {
        "properties": {
          "response_type": {
            "type": "string",
            "enum": [
              "interim_acknowledgment",
              "information_request",
              "final_resolution",
              "referral",
              "repetitive"
            ],
            "title": "Response Type"
          },
          "concerns": {
            "items": {
              "$ref": "#/components/schemas/ConcernResponse"
            },
            "type": "array",
            "title": "Concerns"
          },
          "claims": {
            "items": {
              "$ref": "#/components/schemas/FactualClaim"
            },
            "type": "array",
            "title": "Claims"
          },
          "attachment_ids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "title": "Attachment Ids"
          },
          "request_id": {
            "type": "string",
            "format": "uuid",
            "title": "Request Id"
          },
          "expected_revision": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Expected Revision"
          },
          "actor": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Actor",
            "default": "Demo reviewer"
          },
          "draft_id": {
            "type": "string",
            "format": "uuid",
            "title": "Draft Id"
          },
          "draft_version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Draft Version"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "response_type",
          "concerns",
          "claims",
          "attachment_ids",
          "request_id",
          "expected_revision",
          "draft_id",
          "draft_version"
        ],
        "title": "DraftEdit"
      },
...
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse"
      },
...
      "ToolDefinition": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "system": {
            "type": "string",
            "title": "System"
          },
          "mutates": {
            "type": "boolean",
            "title": "Mutates"
          },
          "simulated": {
            "type": "boolean",
            "const": true,
            "title": "Simulated",
            "default": true
          },
          "parameters": {
            "additionalProperties": true,
            "type": "object",
            "title": "Parameters"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name",
          "system",
          "mutates",
          "parameters"
        ],
        "title": "ToolDefinition"
      },
      "ToolRequest": {
        "properties": {
          "loan_identifier": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Loan Identifier"
          },
          "client_code": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Client Code"
          },
          "command": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/Inspect"
              },
              {
                "$ref": "#/components/schemas/Retrieve"
              },
              {
                "$ref": "#/components/schemas/ApplyPlan"
              },
              {
                "$ref": "#/components/schemas/Prepare"
              },
              {
                "$ref": "#/components/schemas/DraftAction"
              },
              {
                "$ref": "#/components/schemas/CreateTask"
              },
              {
                "$ref": "#/components/schemas/UpdateName"
              },
              {
                "$ref": "#/components/schemas/Reconcile"
              }
            ],
            "title": "Command",
            "discriminator": {
              "propertyName": "operation",
              "mapping": {
                "action.reconcile": "#/components/schemas/Reconcile",
                "action.status": "#/components/schemas/Retrieve",
                "cct.apply_plan": "#/components/schemas/ApplyPlan",
                "cct.close": "#/components/schemas/DraftAction",
                "cct.handoff": "#/components/schemas/ApplyPlan",
                "cct.history": "#/components/schemas/Inspect",
                "cct.read": "#/components/schemas/Inspect",
                "cct.related": "#/components/schemas/Inspect",
                "cct.worklist": "#/components/schemas/Inspect",
                "csp.drafts": "#/components/schemas/Inspect",
                "csp.outbox": "#/components/schemas/Inspect",
                "csp.prepare": "#/components/schemas/Prepare",
                "csp.send": "#/components/schemas/DraftAction",
                "csp.status": "#/components/schemas/Retrieve",
                "ils.create_task": "#/components/schemas/CreateTask",
                "ils.final_note": "#/components/schemas/DraftAction",
                "ils.history": "#/components/schemas/Inspect",
                "ils.read": "#/components/schemas/Inspect",
                "ils.task": "#/components/schemas/Retrieve",
                "ils.tasks": "#/components/schemas/Inspect",
                "ils.update_name": "#/components/schemas/UpdateName",
                "knowledge.search": "#/components/schemas/Inspect",
                "onbase.index": "#/components/schemas/DraftAction",
                "onbase.retrieve": "#/components/schemas/Retrieve",
                "onbase.search": "#/components/schemas/Inspect",
                "onbase.validate": "#/components/schemas/Inspect",
                "review.read": "#/components/schemas/Inspect",
                "rules.assess": "#/components/schemas/Inspect"
              }
            }
          },
          "failure_mode": {
            "type": "string",
            "enum": [
              "none",
              "before_write",
              "after_write_response_lost"
            ],
            "title": "Failure Mode",
            "default": "none"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "loan_identifier",
          "client_code",
          "command"
        ],
        "title": "ToolRequest"
      },
      "ToolResult": {
        "properties": {
          "simulated": {
            "type": "boolean",
            "const": true,
            "title": "Simulated",
            "default": true
          },
          "operation": {
            "type": "string",
            "title": "Operation"
          },
          "status": {
            "type": "string",
            "enum": [
              "ok",
              "prepared",
              "simulated_complete",
              "rejected",
              "awaiting_review",
              "failed",
              "uncertain",
              "in_progress"
            ],
            "title": "Status"
          },
          "action_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Action Id"
          },
          "reference": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ResourceReference"
              },
              {
                "type": "null"
              }
            ]
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "data": {
            "additionalProperties": true,
            "type": "object",
            "title": "Data"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "operation",
          "status",
          "code",
          "message"
        ],
        "title": "ToolResult"
      },
...
      "WorkflowRead": {
        "properties": {
          "case_id": {
            "type": "string",
            "title": "Case Id"
          },
          "revision": {
            "type": "integer",
            "title": "Revision"
          },
          "scenario": {
            "type": "string",
            "title": "Scenario"
          },
          "input_options": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Input Options"
          },
          "pending_work": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Pending Work"
          },
          "drafts": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Drafts"
          },
          "reviews": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Reviews"
          },
          "handoffs": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Handoffs"
          },
          "contacts": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Contacts"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "case_id",
          "revision",
          "scenario",
          "input_options",
          "pending_work",
          "drafts",
          "reviews",
          "handoffs",
          "contacts"
        ],
        "title": "WorkflowRead"
      }
    }
  }
}